+++ client/components/AuthForm.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { authenticate } from "../store";
+import { login } from "../store";
+import { useHistory } from "react-router-dom";
 
 const AuthForm = ({ formName }) => {
   //the only thing we need from the store is the error so we get that using useSelector
@@ -9,7 +10,7 @@ const AuthForm = ({ formName }) => {
       error: state.auth.error,
     };
   });
-
+  let history = useHistory();
   //getting the actions from the store
   const dispatch = useDispatch();
 
@@ -20,7 +21,8 @@ const AuthForm = ({ formName }) => {
   //we need a handle submit function to handle the form submission because of what happens when you submit a form, we need to stop the default behavior of the form which is to refresh the page
   const handleSubmit = (e) => {
     e.preventDefault();
-    dispatch(authenticate(username, password, formName));
+    dispatch(login(username, password));
+    history.push("/home");
   };
 
   return (
@@ -29,7 +31,6 @@ const AuthForm = ({ formName }) => {
         <form
           className="bg-white shadow-lg rounded px-12 pt-6 pb-8 mb-4"
           onSubmit={handleSubmit}
-          name={formName}
         >
           <h2 className="flex justify-center text-center text-2xl font-extrabold py-2 mb-4 text-red-600">
             Login
@@ -59,11 +60,7 @@ const AuthForm = ({ formName }) => {
               className="group relative w-full flex justify-center mt-4 py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-red-600 hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
               type="submit"
             >
-              {formName === "login"
-                ? "Sign In"
-                : formName === "signup"
-                ? "Sign Up"
-                : null}
+              Sign Up
             </button>
           </div>
           {error && error.response && <div> {error.response.data} </div>}
